lazy loading the plugins
